requirements.txt
git+https://github.com/deepghs/waifuc.git@main#egg=waifuc[video]
git+https://github.com/narugo1992/HCP-Diffusion.git@main
dghs-imgutils>=0.3.3
git+https://github.com/narugo1992/gchar.git@main
pillow
numpy
scipy
scikit-learn
huggingface_hub>=0.14.0
requests
click>=7.0.0
hbutils>=0.9.1
di-toolkit
tensorboard
tabulate
markdown2
urlobject
blurhash
dateparser
backports.zoneinfo; python_version < '3.9'
pycivitai>=0.1.2
diffusers>=0.19.0,<0.25
torchsde
natsort
pyquery
pandas
tqdm
torch<2.1.0
pyanimeinfo>=0.0.2
hfutils>=0.0.6
filelock
sdeval>=0.0.3
safetensors